refactor: rename search component
