+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import SomeComponent from './SomeComponent.tsx';
+import SearchComponent from './SearchComponent.tsx';
 import ErrorBoundary from './ErrorBoundary.tsx';
 import './App.css';
 
@@ -7,7 +7,7 @@ export class App extends Component {
   render() {
     return (
       <ErrorBoundary>
-        <SomeComponent />
+        <SearchComponent />
       </ErrorBoundary>
     );
   }
